more cleaner code
